# docs/swagger.yaml
basePath: /
definitions:
  httpserver.dimensions:
    properties:
      height:
        type: integer
      width:
        type: integer
    type: object
  httpserver.metadataResponse:
    properties:
      created_at:
        type: string
      dimensions:
        $ref: '#/definitions/httpserver.dimensions'
      file_name:
        type: string
      id:
        type: string
      mime_type:
        type: string
      size:
        type: integer
      status:
        type: string
      thumbnails:
        items:
          $ref: '#/definitions/httpserver.thumbnail'
        type: array
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  httpserver.thumbnail:
    properties:
      size:
        type: string
      url:
        type: string
    type: object
  httpserver.uploadResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        description: |-
          Status is the public processing state; right after the upload
          it is always "processing".
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: 'Avatar service: uploads, thumbnails and user galleries.'
  title: GophProfile API
  version: "1.0"
paths:
  /api/v1/avatars:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image file (jpeg, png, webp), up to 10 MB
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpserver.uploadResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserID: []
      summary: Upload an avatar
      tags:
      - avatars
  /api/v1/avatars/{avatarID}:
    delete:
      parameters:
      - description: Avatar id
        in: path
        name: avatarID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: deleted
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserID: []
      summary: Delete an avatar
      tags:
      - avatars
    get:
      parameters:
      - description: Avatar id
        in: path
        name: avatarID
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download the original file
      tags:
      - avatars
  /api/v1/avatars/{avatarID}/metadata:
    get:
      parameters:
      - description: Avatar id
        in: path
        name: avatarID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.metadataResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Avatar metadata
      tags:
      - avatars
  /api/v1/avatars/{avatarID}/thumbnails/{size}:
    get:
      parameters:
      - description: Avatar id
        in: path
        name: avatarID
        required: true
        type: string
      - description: Size name
        enum:
        - 100x100
        - 300x300
        in: path
        name: size
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download a thumbnail
      tags:
      - avatars
  /api/v1/users/{userID}/avatar:
    delete:
      parameters:
      - description: User id (must match X-User-ID)
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: deleted
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserID: []
      summary: Delete the user's latest avatar
      tags:
      - users
    get:
      parameters:
      - description: User id
        in: path
        name: userID
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download the user's latest avatar
      tags:
      - users
  /api/v1/users/{userID}/avatars:
    get:
      parameters:
      - description: User id
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpserver.metadataResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the user's avatars
      tags:
      - users
securityDefinitions:
  UserID:
    description: The requester's user id; the MVP trusts it instead of real authentication.
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"